Extract librosa tests from test_transforms to the dedicated test module
